+++ tests/test_example.py
@@ -1,80 +1,145 @@
-"""Perform tests of our examples."""
+"""erform tests of our examples."""
 
 import pytest
 
 import example_client
-from shv import RpcMethodNotFoundError, RpcSubscription, shvmeta_eq
+from shv import RpcMethodAccess, RpcMethodNotFoundError, RpcSubscription, shvmeta
 
 
 @pytest.mark.parametrize(
-    "path,result",
+    "path,method,param,result",
     (
-        ("", [".broker", "test"]),
-        ("test", ["device", "someInt", "someText", "uptime"]),
-        ("test/device", [".app", "track"]),
-        ("test/device/track", [str(i) for i in range(1, 9)]),
-    ),
-)
-async def test_ls(example_device, client, path, result):
-    res = await client.call(path, "ls")
-    assert shvmeta_eq(res, result)
-
-
-@pytest.mark.parametrize(
-    "path,result",
-    (
+        ("", "ls", "test", True),
+        (
+            "test",
+            "ls",
+            None,
+            [
+                "device",
+                "intProperty",
+                "myStructProperty1",
+                "myStructProperty2",
+                "rpcValueProperty",
+                "stringProperty",
+                "uptime",
+                "uptime-duration",
+            ],
+        ),
+        ("test/device", "ls", None, [".app", "track"]),
+        ("test/device/track", "ls", None, [str(i) for i in range(1, 9)]),
+        ("test/device/track", "ls", "1", True),
+        ("test/device/track", "ls", "8", True),
+        ("test/device/track", "ls", "0", False),
+        ("test/device/track", "ls", "9", False),
         (
             "test/device",
+            "dir",
+            None,
             [
-                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: "bws"},
-                {1: "ls", 2: 0, 3: "ils", 4: "ols", 5: "bws"},
-                {1: "lsmod", 2: 1, 4: "olsmod", 5: "bws", 6: "ls"},
+                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: RpcMethodAccess.BROWSE},
+                {
+                    1: "ls",
+                    2: 0,
+                    3: "ils",
+                    4: "ols",
+                    5: RpcMethodAccess.BROWSE,
+                    6: {"lsmod": "olsmod"},
+                },
             ],
         ),
         (
             "test/device/.app",
+            "dir",
+            None,
             [
-                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: "bws"},
-                {1: "ls", 2: 0, 3: "ils", 4: "ols", 5: "bws"},
-                {1: "lsmod", 2: 1, 4: "olsmod", 5: "bws", 6: "ls"},
-                {1: "shvVersionMajor", 2: 2, 4: "Int", 5: "rd"},
-                {1: "shvVersionMinor", 2: 2, 4: "Int", 5: "rd"},
-                {1: "name", 2: 2, 4: "String", 5: "rd"},
-                {1: "version", 2: 2, 4: "String", 5: "rd"},
-                {1: "date", 2: 2, 4: "DateTime", 5: "rd"},
-                {1: "ping", 2: 0, 5: "bws"},
+                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: RpcMethodAccess.BROWSE},
+                {
+                    1: "ls",
+                    2: 0,
+                    3: "ils",
+                    4: "ols",
+                    5: RpcMethodAccess.BROWSE,
+                    6: {"lsmod": "olsmod"},
+                },
+                {1: "shvVersionMajor", 2: 2, 4: "Int", 5: RpcMethodAccess.READ},
+                {1: "shvVersionMinor", 2: 2, 4: "Int", 5: RpcMethodAccess.READ},
+                {1: "name", 2: 2, 4: "String", 5: RpcMethodAccess.READ},
+                {1: "version", 2: 2, 4: "String", 5: RpcMethodAccess.READ},
+                {1: "date", 2: 2, 4: "DateTime", 5: RpcMethodAccess.READ},
+                {1: "ping", 2: 0, 5: RpcMethodAccess.BROWSE},
             ],
         ),
         (
             "test/device/track",
+            "dir",
+            None,
             [
-                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: "bws"},
-                {1: "ls", 2: 0, 3: "ils", 4: "ols", 5: "bws"},
-                {1: "lsmod", 2: 1, 4: "olsmod", 5: "bws", 6: "ls"},
-                {1: "reset", 2: 32, 5: "cmd"},
-                {1: "lastResetUser", 2: 2, 3: "Int", 4: "StringOrNull", 5: "rd"},
+                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: RpcMethodAccess.BROWSE},
+                {
+                    1: "ls",
+                    2: 0,
+                    3: "ils",
+                    4: "ols",
+                    5: RpcMethodAccess.BROWSE,
+                    6: {"lsmod": "olsmod"},
+                },
+                {
+                    1: "reset",
+                    2: 32,
+                    5: RpcMethodAccess.COMMAND,
+                    63: {"dependencies": "Reset all tracks to their initial state"},
+                },
+                {
+                    1: "lastResetUser",
+                    2: 2,
+                    3: "Int",
+                    4: "StringOrNull",
+                    5: RpcMethodAccess.READ,
+                },
             ],
         ),
         (
             "test/device/track/1",
+            "dir",
+            None,
             [
-                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: "bws"},
-                {1: "ls", 2: 0, 3: "ils", 4: "ols", 5: "bws"},
-                {1: "lsmod", 2: 1, 4: "olsmod", 5: "bws", 6: "ls"},
-                {1: "get", 2: 2, 3: "Int", 4: "List[Int]", 5: "rd"},
-                {1: "set", 2: 4, 3: "List[Int]", 5: "wr"},
+                {1: "dir", 2: 0, 3: "idir", 4: "odir", 5: RpcMethodAccess.BROWSE},
+                {
+                    1: "ls",
+                    2: 0,
+                    3: "ils",
+                    4: "ols",
+                    5: RpcMethodAccess.BROWSE,
+                    6: {"lsmod": "olsmod"},
+                },
+                {
+                    1: "get",
+                    2: 2,
+                    3: "Int",
+                    4: "List[Int]",
+                    5: RpcMethodAccess.READ,
+                    6: {"chng": "List[Int]"},
+                    63: {"description": "List of tracks"},
+                },
+                {
+                    1: "set",
+                    2: 4,
+                    3: "List[Int]",
+                    5: RpcMethodAccess.WRITE,
+                    63: {"description": "Set track"},
+                },
             ],
         ),
+        ("test/device/track/4", "dir", "get", True),
+        ("test/device/track/4", "dir", "set", True),
+        ("test/device/track/4", "dir", "chng", False),
+        ("test/device/track/4", "get", None, [0, 1, 2, 3]),
     ),
 )
-async def test_dir(example_device, client, path, result):
-    res = await client.call(path, "dir")
-    assert shvmeta_eq(res, result)
-
-
-async def test_get(example_device, client):
-    res = await client.call("test/device/track/4", "get")
-    assert res == [0, 1, 2, 3]
+async def test_call(example_device, client, path, method, param, result):
+    res = await client.call(path, method, param)
+    assert res == result
+    assert shvmeta(res) == shvmeta(result)
 
 
 async def test_set(example_device, client):
@@ -82,7 +147,8 @@ async def test_set(example_device, client):
     assert await client.call("test/device/track/4", "set", tracks) is None
 
     res = await client.call("test/device/track/4", "get")
-    assert shvmeta_eq(res, tracks)
+    assert res == tracks
+    assert shvmeta(res) == shvmeta(tracks)
 
 
 async def test_reset(example_device, value_client):